the beginnings of combat
Also level geometry for first area of the game

+++ DiletsAdventure/src/Coordinate.java
@@ -1,20 +1,20 @@
 public class Coordinate {
-    int x;
-    int y;
-    public Coordinate(int x, int y){
+    double x;
+    double y;
+    public Coordinate(double x, double y){
         this.x = x;
         this.y = y;
     }
-    public int getX(){
+    public double getX(){
         return this.x;
     }
-    public int getY(){
+    public double getY(){
         return this.y;
     }
-    public void addX(int a){
+    public void addX(double a){
         this.x+=a;
     }
-    public void addY(int a){
+    public void addY(double a){
         this.y+=a;
     }
